chat/requirements.txt
PyQt6==6.7.0
paho-mqtt==2.1.0
